Split RxUtilsTest into two differnt Test classes
